refactor: move add repository API from ProjectController to RepositoryController
