+++ src/main/java/pvs/app/controller/ProjectController.java
@@ -5,9 +5,9 @@ import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
-import pvs.app.dto.*;
+import pvs.app.dto.CreateProjectDTO;
+import pvs.app.dto.ResponseProjectDTO;
 import pvs.app.service.ProjectService;
-import pvs.app.service.RepositoryService;
 
 import java.util.List;
 import java.util.Optional;
@@ -16,88 +16,19 @@ import java.util.Optional;
 @RequestMapping(produces = MediaType.APPLICATION_JSON_VALUE)
 public class ProjectController {
     private final ProjectService projectService;
-    private final RepositoryService repositoryService;
-    @Value("${message.exception}")
-    private String exceptionMessage;
-    @Value("${message.invalid.url}")
-    private String urlInvalidMessage;
     @Value("${message.success}")
     private String successMessage;
     @Value("${message.fail}")
     private String failMessage;
 
-    public ProjectController(ProjectService projectService, RepositoryService repositoryService) {
+    public ProjectController(ProjectService projectService) {
         this.projectService = projectService;
-        this.repositoryService = repositoryService;
     }
 
     @PostMapping("/project")
     public ResponseEntity<String> createProject(@RequestBody CreateProjectDTO createProjectDTO) {
         projectService.create(createProjectDTO);
         return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-    }
-
-    @PostMapping("/project/{projectId}/repository/sonar")
-    public ResponseEntity<String> addSonarRepository(@RequestBody AddRepositoryDTO addSonarRepositoryDTO) {
-        try {
-            if (repositoryService.checkSonarURL(addSonarRepositoryDTO.getRepositoryURL())) {
-                if (projectService.addSonarRepo(addSonarRepositoryDTO)) {
-                    return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-                }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(failMessage);
-            }
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-        } catch (Exception e) {
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exceptionMessage);
-        }
-    }
-
-    @PostMapping("/project/{projectId}/repository/github")
-    public ResponseEntity<String> addGitHubRepository(@RequestBody AddRepositoryDTO addRepositoryDTO) {
-        try {
-            if (repositoryService.checkGithubURL(addRepositoryDTO.getRepositoryURL())) {
-                if (projectService.addGithubRepo(addRepositoryDTO)) {
-                    return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-                }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(failMessage);
-            }
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exceptionMessage);
-        }
-    }
-
-    @PostMapping("/project/{projectId}/repository/gitlab")
-    public ResponseEntity<String> addGitLabRepository(@RequestBody AddRepositoryDTO addGitLabRepositoryDTO) {
-        try {
-            if (repositoryService.checkGitLabURL(addGitLabRepositoryDTO.getRepositoryURL())) {
-                if (projectService.addGitLabRepo(addGitLabRepositoryDTO)) {
-                    return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-                }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(failMessage);
-            }
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-        } catch (Exception e) {
-            System.out.println("@@@@@@@@@@@@@@@@@@@@@@@@");
-            e.printStackTrace();
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exceptionMessage);
-        }
-    }
-
-    @PostMapping("/project/{projectId}/repository/trello")
-    public ResponseEntity<String> addTrelloBoard(@RequestBody AddRepositoryDTO addTrelloBoardDTO) {
-        try {
-            if (repositoryService.checkTrelloURL(addTrelloBoardDTO.getRepositoryURL())) {
-                if (projectService.addTrelloBoard(addTrelloBoardDTO)) {
-                    return ResponseEntity.status(HttpStatus.OK).body(successMessage);
-                }
-                return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(failMessage);
-            }
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(urlInvalidMessage);
-        } catch (Exception e) {
-            return ResponseEntity.status(HttpStatus.BAD_REQUEST).body(exceptionMessage);
-        }
     }
 
     @GetMapping("/project/{memberId}")
